requirements.txt
# Core Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# AI and Machine Learning
google-genai==0.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4

# Audio and Video Processing
SpeechRecognition==3.10.0
# Fixed OpenCV and NumPy compatibility
numpy==1.24.3
opencv-python==4.8.1.78
Pillow>=11.0.0
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0

# Data Processing and Utilities
pandas>=2.0.0
python-dotenv==1.0.0

# HTTP and API
requests==2.31.0
urllib3==2.0.7

# Development and Testing
pytest==7.4.2
pytest-cov==4.1.0

# System Utilities
psutil==5.9.5
